Revert warning duration type from short to int and refactor showClusters according to Chris's suggestion

+++ zigbee-api/src/main/java/org/bubblecloud/zigbee/api/cluster/impl/api/security_safety/ias_wd/StartWarningPayload.java
@@ -32,5 +32,5 @@ public interface StartWarningPayload {
 
     public short getStrobe();
 
-    public short getWarningDuration();
+    public int getWarningDuration();
 }
